Time and Time Zones
The app now subscribes to the Time and will fetch the Time Zone on initialisation. However, there are no safety mechanisms for initialisation yet.

The Time and Time Zone are stored in Sheet, even though they won't actually get stored in JSON. (This may require data structure changes in the future but I'm keeping it as simple as possible until it becomes obviously necessary.)

- New Sheet.updateTime.
- New Sheet.updateTimeZone.
- Sheet.startCurrentBlock is now complete - it uses the current time to add the start time to the block.
- Sheet.endCurrentBlock is now functionally complete - it uses the current time to add the end time to the block.