New mass edit, query and helper functions
Started on functions for mass editing blocks, starting with one that can find and replace subtask IDs for a certain ProjectType.
- New Sheet.replaceSubtaskIDs.
- New Project.replaceSubtaskIDs.

Another new statistic function for Project - toTotalMoney, to calculate total money earned for a project (important if it's work on an hourly rate)
- New Project.toTotalMoney.

New query function for Project, to keep Project as encapsulated as possible.
- New Project.hasProjectTypeID.

New Helper module to centralise common functions
- New Helper.millisToHours to convert millisecond time values to hours.